// src/app/providers/navigation/types.ts
import {NativeStackScreenProps} from '@react-navigation/native-stack';

export type RootScreenProps<
  S extends keyof RootStackParamList = keyof RootStackParamList,
> = NativeStackScreenProps<RootStackParamList, S>;

export type RootStackParamList = {
  Login: undefined;
  Splash: undefined;
  Checkin: undefined;
  ChooseShop: undefined;
  Settings: undefined;
  EditProfileUser: undefined;
  System: undefined;
  Report: undefined;
  BottomNavigator: undefined;
  BookingManage: undefined;
  DetailBookingItem: { bookingId: number | string };
  AddNewBooking: undefined;
  EditBooking: undefined;
  DashBoard: undefined;
};